update: tests

+++ test/GaugedERC20.t.sol
@@ -11,8 +11,8 @@ contract GaugedUniERC20Test is Test, Gauge {
     GaugedUniERC20 public dstToken__;
     MockSocket public mockSocket__;
 
-    uint256 chainSlug_ = 1;
-    uint256 siblingChainSlug_ = 2;
+    uint32 chainSlug_ = 1;
+    uint32 siblingChainSlug_ = 2;
 
     address public constant fastSwitchboard = address(1);
     address public constant optimisticSwitchboard = address(2);
